app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.30.0
psycopg[binary]>=3.2.4
alembic>=1.13.0
pydantic>=2.10
pydantic-settings>=2.0
python-multipart>=0.0.20
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
PyJWT>=2.10.1
email-validator>=2.2.0
bcrypt==4.1.2
httpx>=0.28.0
